Refactor assetRedirectMixin in screen.vue and ScriptEditor.vue
