+++ resources/js/modules/assetRedirectMixin.js
@@ -1,4 +1,15 @@
 export default {
+  data() {
+    return {
+      redirectUrl: null,
+    };
+  },
+  mounted() {
+    this.setRedirectUrl();
+    ProcessMaker.EventBus.$on("redirect", () => {
+      this.handleRedirection();
+    });
+  },
   computed: {
     isProjectRoute() {
       return window.location.pathname.startsWith("/designer/projects/");
@@ -8,6 +19,19 @@ export default {
     appendProjectIdToURL(url, projectId) {
       if (this.isProjectRoute) {
         url.searchParams.append("project_id", projectId);
+      }
+    },
+    setRedirectUrl() {
+      const queryParams = new URLSearchParams(window.location.search);
+      const projectId = queryParams.get("project_id");
+
+      if (projectId) {
+        this.redirectUrl = `/designer/projects/${projectId}`;
+      }
+    },
+    handleRedirection() {
+      if (this.redirectUrl) {
+        window.location.href = this.redirectUrl;
       }
     },
   },
